// React/app/GerAdv_TS/Auditor4K/Interfaces/interface.Auditor4K.ts
// 100% auto-generated by Source Genesys WebApi Front & Back Creator
                     // copyright © 2000-2025 Menphis - Sistemas Inteligentes
                     // This file is part of the Source Genesys project                     
                    'use client';

export interface IAuditor4K {
// 202501251
    id: number;
 
	nome: string,
}

export interface IAuditor4KFormProps {
  id: number;
  onClose: () => void;
  onSuccess: (registro?: any) => void;
  onError: () => void;
}

export interface IAuditor4KIncProps {
    id: number;
    onClose: () => void;
    onSuccess: (registro?: any) => void;
    onError: () => void;
}